# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PerformanceOS AI - Marketing Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/kpis:
    get:
      operationId: getKpis
      tags: [dashboard]
      summary: Get KPI metrics with trends
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, yesterday, last7days, last30days, last90days]
            default: last30days
      responses:
        "200":
          description: KPI metrics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KpiCard"
  /dashboard/platform-comparison:
    get:
      operationId: getPlatformComparison
      tags: [dashboard]
      summary: Compare performance across ad platforms
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [last7days, last30days, last90days]
            default: last30days
      responses:
        "200":
          description: Platform comparison data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatformStat"
  /dashboard/spend-trend:
    get:
      operationId: getSpendTrend
      tags: [dashboard]
      summary: Get daily spend trend chart data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [last7days, last30days, last90days]
            default: last30days
      responses:
        "200":
          description: Spend trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpendTrendPoint"
...
components:
  schemas:
...
    KpiCard:
      type: object
      required: [id, label, value, unit, trend, trendValue, sparkline]
      properties:
        id:
          type: string
        label:
          type: string
        value:
          type: number
        unit:
          type: string
        trend:
          type: string
          enum: [up, down, neutral]
        trendValue:
          type: number
        sparkline:
          type: array
          items:
            type: number
        description:
          type: string
    PlatformStat:
      type: object
      required: [platform, spend, leads, roas, cpa, revenue, color]
      properties:
        platform:
          type: string
          enum: [google, meta, linkedin, microsoft]
        spend:
          type: number
        leads:
          type: integer
        roas:
          type: number
        cpa:
          type: number
        revenue:
          type: number
        ctr:
          type: number
        cpc:
          type: number
        color:
          type: string
    SpendTrendPoint:
      type: object
      required: [date, google, meta, linkedin, microsoft, total]
      properties:
        date:
          type: string
        google:
          type: number
        meta:
          type: number
        linkedin:
          type: number
        microsoft:
          type: number
        total:
          type: number
...